requirements.txt
beautifulsoup4==4.13.3
Flask==3.1.0
openpyxl==3.1.5
Pillow==11.1.0
pymongo==4.11
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.1.2
requests==2.32.3
sentence-transformers==3.4.1
torch==2.5.1
Werkzeug==3.1.3
gunicorn==20.1.0